[#1349] Use `--scalar-values` option if available.
Newer versions of GDB support the `--scalar-values` option (with value
3) to the `-stack-list-arguments`, `-stack-list-locals` and
`-stack-list-variables` commands. This option fetches names, types,
and values for scalar types, or names and types only for non-scalar
types.

Use this in `DebuggedProcess.GetParameterInfoOnly()` if it is
available, as this gets all the required information without the
potential performance penalty of the `--simple-values` option which
potentially fetches large values for references to arrays, structures,
and unions.

New method `MICommandFactory.SupportsScalarValues()` determines if
`--scalar-values` is supported, using the `-list-features` command and
checking for the presence of the `"scalar-values"` feature. We cache
the result on the `DebuggedProcess` object, as the set of supported
features does not change during the lifetime of the debug session.